apps/asr-engine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
requests>=2.31
numpy<2.0
soundfile
librosa>=0.10.1
ctranslate2>=4.3
faster-whisper>=1.0
pyannote.audio==3.1.1
